requirements.txt
boto3==1.34.162
clickhouse-connect==0.7.16
cryptography==42.0.8
extra-streamlit-components==0.1.71
fastapi==0.110.2
pandas==2.2.2
psycopg2-binary==2.9.9
pyarrow==17.0.0
python-dateutil==2.9.0.post0
python-multipart==0.0.9
PyYAML==6.0.2
requests==2.32.3
streamlit==1.37.1
uvicorn==0.30.1
